更换依赖注入框架dibus,事件分发采用dibus

+++ main/src/main/java/com/example/main/http/entity/DiscoveryEntity.kt
@@ -52,7 +52,7 @@ data class Resource(
 
 data class ResourceExtInfo(
     val highQuality: Boolean,
-    val playCount: Int
+    val playCount: Long
 )
 
 data class UiElement(
